+++ backend/windmill-common/src/error.rs
@@ -90,7 +90,7 @@ impl IntoResponse for Error {
         if matches!(status, axum::http::StatusCode::NOT_FOUND) {
             tracing::warn!(message = e.to_string());
         } else {
-            tracing::error!(message = e.to_string());
+            tracing::error!(message = e.to_string(), error = ?e);
         };
 
         axum::response::Response::builder()