+++ frontend/src/lib/components/LogViewer.svelte
@@ -61,7 +61,7 @@
 					icon: Download
 				}}
 			>
-				Downlaod
+				Download
 			</Button>
 
 			<Button