+++ frontend/src/lib/components/ScriptBuilder.svelte
@@ -108,9 +108,9 @@
 	}
 
 	const langs: [string, SupportedLanguage][] = [
-		['TypeScript (Deno)', Script.language.DENO],
+		['TypeScript (Bun)', Script.language.BUN],
 		['Python', Script.language.PYTHON3],
-		['TypeScript (Bun)', Script.language.BUN]
+		['TypeScript (Deno)', Script.language.DENO]
 	]
 	if (SCRIPT_SHOW_BASH) {
 		langs.push(['Bash', Script.language.BASH])
@@ -686,8 +686,13 @@
 												type="text"
 												autofocus
 												bind:value={script.concurrency_key}
-												placeholder={`$workspace/script/${script.path}`}
+												placeholder={`$workspace/script/${script.path}-$args[foo]`}
 											/>
+											<Tooltip
+												>Concurrency keys are global, you can have them be workspace specific using
+												the variable `$workspace`. You can also use an argument's value using
+												`$args[name_of_arg]`</Tooltip
+											>
 										</Label>
 									</div>
 								</Section>