+++ frontend/src/lib/components/apps/components/layout/AppDecisionTree.svelte
@@ -151,6 +151,21 @@
 			subGridIndex: nodes.findIndex((node) => node.id === currentNodeId)
 		}
 	}
+
+	function updateDebuggingComponents() {
+		const nodeIds = nodes.map((node) => node.id).sort()
+		const debuggingComponentsIds = Object.keys($debuggingComponents).sort()
+
+		if (JSON.stringify(nodeIds) !== JSON.stringify(debuggingComponentsIds)) {
+			debuggingComponentsIds.forEach((id) => {
+				if (!nodeIds.includes(id)) {
+					delete $debuggingComponents[id]
+				}
+			})
+		}
+	}
+
+	$: nodes && $debuggingComponents && updateDebuggingComponents()
 </script>
 
 {#if Object.keys(resolvedConditions).length === nodes.length}