+++ frontend/src/lib/components/flows/content/FlowSettings.svelte
@@ -493,6 +493,11 @@
 										bind:seconds={$flowStore.value.concurrency_time_window_s}
 									/>
 								</Label>
+								<Label label="Custom concurrency key">
+									<div class="text-tertiary text-xs"
+										>Custom concurrency keys can only be set as the setting of a workspace script</div
+									>
+								</Label>
 							</div>
 						</Section>
 					</TabContent>